move setting to general tab
