+++ packages/yoroi-extension/app/containers/settings/categories/BlockchainSettingsPage.js
@@ -3,17 +3,13 @@ import type { Node } from 'react';
 import { Component } from 'react';
 import { observer } from 'mobx-react';
 import { computed } from 'mobx';
-import { defineMessages, intlShape } from 'react-intl';
+import { intlShape } from 'react-intl';
 import ExplorerSettings from '../../../components/settings/categories/general-setting/ExplorerSettings';
-import UnitOfAccountSettings from '../../../components/settings/categories/general-setting/UnitOfAccountSettings';
 import type { InjectedOrGenerated } from '../../../types/injectedPropsType';
 import UriSettingsBlock from '../../../components/settings/categories/general-setting/UriSettingsBlock';
 import registerProtocols from '../../../uri-protocols';
 import environment from '../../../environment';
-import { unitOfAccountDisabledValue } from '../../../types/unitOfAccountType';
-import { ReactComponent as AdaCurrency }  from '../../../assets/images/currencies/ADA.inline.svg';
 import type { $npm$ReactIntl$IntlFormat } from 'react-intl';
-import type { UnitOfAccountSettingType } from '../../../types/unitOfAccountType';
 import LocalizableError from '../../../i18n/LocalizableError';
 import {
   PublicDeriver,
@@ -28,41 +24,6 @@ import type {
   GetAllExplorersResponse,
 } from '../../../api/ada/lib/storage/bridge/explorers';
 
-const currencyLabels = defineMessages({
-  USD: {
-    id: 'settings.unitOfAccount.currency.usd',
-    defaultMessage: '!!!US dollar',
-  },
-  JPY: {
-    id: 'settings.unitOfAccount.currency.jpy',
-    defaultMessage: '!!!Japanese yen',
-  },
-  EUR: {
-    id: 'settings.unitOfAccount.currency.eur',
-    defaultMessage: '!!!Euro',
-  },
-  CNY: {
-    id: 'settings.unitOfAccount.currency.cny',
-    defaultMessage: '!!!Chinese Renminbi yuan',
-  },
-  KRW: {
-    id: 'settings.unitOfAccount.currency.krw',
-    defaultMessage: '!!!South Korean won',
-  },
-  BTC: {
-    id: 'settings.unitOfAccount.currency.btc',
-    defaultMessage: '!!!Bitcoin',
-  },
-  ETH: {
-    id: 'settings.unitOfAccount.currency.eth',
-    defaultMessage: '!!!Ethereum',
-  },
-  BRL: {
-    id: 'settings.unitOfAccount.currency.brl',
-    defaultMessage: '!!!Brazilian real',
-  },
-});
-
 type GeneratedData = typeof BlockchainSettingsPage.prototype.generated;
 
 @observer
@@ -70,13 +31,6 @@ export default class BlockchainSettingsPage extends Component<InjectedOrGenerate
 
   static contextTypes: {|intl: $npm$ReactIntl$IntlFormat|} = {
     intl: intlShape.isRequired,
-  };
-
-  onSelectUnitOfAccount: string => Promise<void> = async (value) => {
-    const unitOfAccount = (value === 'ADA')
-      ? unitOfAccountDisabledValue
-      : { enabled: true, currency: value };
-    await this.generated.actions.profile.updateUnitOfAccount.trigger(unitOfAccount);
   };
 
   render(): Node {
@@ -89,8 +43,6 @@ export default class BlockchainSettingsPage extends Component<InjectedOrGenerate
     const { stores } = this.generated;
 
     const isSubmittingExplorer = stores.explorers.setSelectedExplorerRequest.isExecuting;
-    const isSubmittingUnitOfAccount = stores.profile.setUnitOfAccountRequest.isExecuting
-      || stores.coinPriceStore.refreshCurrentUnit.isExecuting;
 
     const uriSettings = (
       isCardanoHaskell(networkInfo) &&
@@ -103,28 +55,6 @@ export default class BlockchainSettingsPage extends Component<InjectedOrGenerate
         />
       )
       : null;
-
-    const currencies = stores.profile.UNIT_OF_ACCOUNT_OPTIONS.map(c => {
-      const name = this.context.intl.formatMessage(currencyLabels[c.symbol]);
-      return {
-        value: c.symbol,
-        label: `${c.symbol} - ${name}`,
-        name,
-        price: stores.coinPriceStore.getCurrentPrice('ADA', c.symbol),
-        svg: c.svg
-      };
-    });
-    currencies.unshift({
-      value: 'ADA',
-      label: 'ADA - Cardano',
-      name: 'Cardano',
-      native: true,
-      svg: AdaCurrency,
-    });
-
-    const unitOfAccountValue = stores.profile.unitOfAccount.enabled
-      ? stores.profile.unitOfAccount.currency
-      : 'ADA';
 
     return (
       <>
@@ -140,14 +70,6 @@ export default class BlockchainSettingsPage extends Component<InjectedOrGenerate
           error={stores.explorers.setSelectedExplorerRequest.error}
         />
         {uriSettings}
-        <UnitOfAccountSettings
-          onSelect={this.onSelectUnitOfAccount}
-          isSubmitting={isSubmittingUnitOfAccount}
-          currencies={currencies}
-          currentValue={unitOfAccountValue}
-          error={stores.profile.setUnitOfAccountRequest.error}
-          lastUpdatedTimestamp={stores.coinPriceStore.lastUpdateTimestamp}
-        />
       </>
     );
   }
@@ -161,24 +83,9 @@ export default class BlockchainSettingsPage extends Component<InjectedOrGenerate
           |}) => Promise<void>
         |},
       |},
-      profile: {|
-        updateUnitOfAccount: {|
-          trigger: (
-            params: UnitOfAccountSettingType
-          ) => Promise<void>
-        |}
-      |}
     |},
     canRegisterProtocol: () => boolean,
     stores: {|
-      coinPriceStore: {|
-        getCurrentPrice: (
-          from: string,
-          to: string
-        ) => ?string,
-        lastUpdateTimestamp: null | number,
-        refreshCurrentUnit: {| isExecuting: boolean |}
-      |},
       explorers: {|
         selectedExplorer: Map<number, SelectedExplorer>,
         setSelectedExplorerRequest: {|
@@ -186,17 +93,6 @@ export default class BlockchainSettingsPage extends Component<InjectedOrGenerate
           isExecuting: boolean
         |},
         allExplorers: GetAllExplorersResponse,
-      |},
-      profile: {|
-        UNIT_OF_ACCOUNT_OPTIONS: Array<{|
-          svg: string,
-          symbol: string
-        |}>,
-        setUnitOfAccountRequest: {|
-          error: ?LocalizableError,
-          isExecuting: boolean
-        |},
-        unitOfAccount: UnitOfAccountSettingType
       |},
       wallets: {|
         selected: null | PublicDeriver<>
@@ -221,21 +117,6 @@ export default class BlockchainSettingsPage extends Component<InjectedOrGenerate
             error: stores.explorers.setSelectedExplorerRequest.error,
           },
         },
-        profile: {
-          UNIT_OF_ACCOUNT_OPTIONS: profileStore.UNIT_OF_ACCOUNT_OPTIONS,
-          unitOfAccount: profileStore.unitOfAccount,
-          setUnitOfAccountRequest: {
-            error: profileStore.setUnitOfAccountRequest.error,
-            isExecuting: profileStore.setUnitOfAccountRequest.isExecuting,
-          },
-        },
-        coinPriceStore: {
-          getCurrentPrice: stores.coinPriceStore.getCurrentPrice,
-          lastUpdateTimestamp: stores.coinPriceStore.lastUpdateTimestamp,
-          refreshCurrentUnit: {
-            isExecuting: stores.coinPriceStore.refreshCurrentUnit.isExecuting,
-          },
-        },
         wallets: {
           selected: stores.wallets.selected,
         },
@@ -243,9 +124,6 @@ export default class BlockchainSettingsPage extends Component<InjectedOrGenerate
       actions: {
         explorers: {
           updateSelectedExplorer: { trigger: actions.explorers.updateSelectedExplorer.trigger },
-        },
-        profile: {
-          updateUnitOfAccount: { trigger: actions.profile.updateUnitOfAccount.trigger },
         },
       },
       canRegisterProtocol: environment.userAgentInfo.canRegisterProtocol,
