+++ packages/yoroi-extension/app/containers/settings/categories/GeneralSettingsPage.stories.js
@@ -12,6 +12,7 @@ import { walletLookup } from '../../../../stories/helpers/WalletCache';
 import { wrapSettings } from '../../../Routes';
 import { mockSettingsProps } from '../Settings.mock';
 import { ROUTES } from '../../../routes-config';
+import { SUPPORTED_CURRENCIES } from '../../../config/unitOfAccount';
 
 export default {
   title: `${__filename.split('.')[0]}`,
@@ -41,6 +42,23 @@ export const Generic = (): Node => {
             currentLocale: globalKnobs.locale(),
             currentTheme: globalKnobs.currentTheme(),
             hasCustomTheme: () => boolean('hasCustomTheme', false),
+
+            UNIT_OF_ACCOUNT_OPTIONS: SUPPORTED_CURRENCIES,
+            unitOfAccount: {
+              enabled: false,
+              currency: undefined,
+            },
+            setUnitOfAccountRequest: {
+              error: null,
+              isExecuting: boolean('setUnitOfAccountRequest_isExecuting'),
+            },
+            coinPriceStore: {
+              getCurrentPrice: (_from, _to) => '5',
+              lastUpdateTimestamp: Date.now(),
+              refreshCurrentUnit: {
+                isExecuting: false,
+              },
+            },
           },
         },
         actions: {
@@ -48,6 +66,7 @@ export const Generic = (): Node => {
             updateLocale: { trigger: async req => action('updateLocale')(req) },
             updateTheme: { trigger: async req => action('updateTheme')(req) },
             exportTheme: { trigger: async req => action('exportTheme')(req) },
+            updateUnitOfAccount: { trigger: async (req) => action('updateUnitOfAccount')(req) },
           },
         },
       }}
